Se termina la rama de administración y gestión de las cuentas

+++ src/app/pages/clients/clients.page.ts
@@ -1,3 +1,4 @@
+import { StorageService } from './../../services/storage.service';
 import { Component, inject, OnInit } from '@angular/core';
 import { orderBy, where } from 'firebase/firestore';
 import { User } from 'src/app/models/user.model';
@@ -14,6 +15,7 @@ export class ClientsPage {
 
   utilsSrv = inject(UtilsService)
   firebaseSrv = inject(FirebaseService)
+  storageSrv = inject(StorageService)
 
   users: User[] = []
   filteredUsers: User[]=[]
@@ -88,15 +90,21 @@ export class ClientsPage {
   }
 
   async deleteClient(user: User) {
-    let path = `users/${user.uid}`;
+    let path = `users/${user.uid}`;  // Ruta del documento del usuario
 
     const loading = await this.utilsSrv.loading();
     await loading.present();
 
     try {
-      let pathImage = await this.firebaseSrv.getFilePath(user.image)
-      await this.firebaseSrv.deleteDocument(path);
-      await this.firebaseSrv.deleteFile(pathImage)
+      if (user.image && user.image.toString() !=="https://firebasestorage.googleapis.com/v0/b/citas-barber-28dfc.appspot.com/o/user_2078.webp?alt=media&token=896ae797-9229-44f0-8b39-27e5da7cac7f") {
+        let pathImage = await this.firebaseSrv.getFilePath(user.image);  // Obtener la ruta de la imagen
+        await this.firebaseSrv.deleteFile(pathImage)
+      }
+
+      // Ahora, eliminamos el documento de Firestore
+      await this.firebaseSrv.deleteDocument(path);  // Eliminar el documento del usuario
+
+      // Mostrar mensaje de éxito
       this.utilsSrv.showToast({
         message: 'Cliente eliminado exitosamente',
         duration: 1500,
@@ -104,15 +112,22 @@ export class ClientsPage {
         position: 'middle',
         icon: 'checkmark-circle-outline',
       });
+
+      // Opcional: Actualizar la lista de clientes después de la eliminación
+      this.getClients();
+
     } catch (error) {
+      // Si ocurre un error, mostramos el mensaje de error
       this.utilsSrv.showToast({
-        message:"Ha ocurrido un error",
+        message: "Ha ocurrido un error al eliminar al cliente",
         duration: 2500,
         color: 'primary',
         position: 'middle',
         icon: 'alert-circle-outline',
       });
+      console.error("Error al eliminar cliente:", error);  // Log del error para depuración
     } finally {
+      // Asegurarse de cerrar el loading, independientemente de si hubo éxito o error
       loading.dismiss();
     }
   }
